fix(chat-tools): reject protocol-relative video URLs; key-based avatar reset; rename platformVisual
- Sora2VideoResult: safeVideoUrl now rejects "//host/…" (protocol-relative)
  URLs that previously passed the startsWith("/") check and could resolve to a
  foreign origin.
- ArtistHeroSection: extract ArtistAvatar and remount it via key={imageUrl}
  instead of resetting load/error state in a useEffect — removes stale-state
  flicker on URL change.
- Rename platformVisual.ts -> getPlatformVisual.ts so the file basename matches
  its single exported function.
