+++ components/VercelChat/tools/ArtistHeroSection.tsx
@@ -6,6 +6,44 @@ import { formatDate } from "date-fns";
 import { ArtistProfile } from "@/lib/supabase/artist/updateArtistProfile";
 
 /**
+ * Avatar with a fade-in load + graceful icon fallback. Its load/error state is
+ * owned locally, and the parent remounts it via a `key={imageUrl}` so a new
+ * avatar starts fresh — no useEffect reset, no stale-state flicker between URLs.
+ */
+const ArtistAvatar = ({ imageUrl, name }: { imageUrl: string; name: string }) => {
+  const [imgLoaded, setImgLoaded] = React.useState(false);
+  const [imgErrored, setImgErrored] = React.useState(false);
+
+  if (imgErrored) {
+    return (
+      <div className="flex size-16 shrink-0 items-center justify-center rounded-xl bg-muted text-muted-foreground shadow-sm">
+        <User className="size-6 opacity-60" />
+      </div>
+    );
+  }
+
+  return (
+    <div className="relative size-16 shrink-0 overflow-hidden rounded-xl bg-muted shadow-sm">
+      {!imgLoaded ? (
+        <div className="absolute inset-0 flex animate-pulse items-center justify-center bg-muted text-muted-foreground">
+          <User className="size-6 opacity-50" />
+        </div>
+      ) : null}
+      {/* eslint-disable-next-line @next/next/no-img-element */}
+      <img
+        src={imageUrl}
+        alt={name}
+        className={`size-full object-cover transition-opacity duration-300 ${
+          imgLoaded ? "opacity-100" : "opacity-0"
+        }`}
+        onLoad={() => setImgLoaded(true)}
+        onError={() => setImgErrored(true)}
+      />
+    </div>
+  );
+};
+
+/**
  * Compact, token-safe artist identity header used inside success tool cards.
  * Shows the artist avatar, name, label and last-updated. The success line only
  * renders when a `message` is passed, so a parent ToolCard that already owns the
@@ -19,36 +57,15 @@ const ArtistHeroSection = ({
   message?: string;
 }) => {
   const imageUrl = artistProfile?.image;
-  const [imgLoaded, setImgLoaded] = React.useState(false);
-  const [imgErrored, setImgErrored] = React.useState(false);
-
-  // Reset load/error state whenever the image URL changes so a new avatar
-  // re-runs the fade-in instead of inheriting the previous image's state.
-  React.useEffect(() => {
-    setImgLoaded(false);
-    setImgErrored(false);
-  }, [imageUrl]);
 
   return (
     <div className="flex items-center gap-4 p-4">
-      {imageUrl && !imgErrored ? (
-        <div className="relative size-16 shrink-0 overflow-hidden rounded-xl bg-muted shadow-sm">
-          {!imgLoaded ? (
-            <div className="absolute inset-0 flex animate-pulse items-center justify-center bg-muted text-muted-foreground">
-              <User className="size-6 opacity-50" />
-            </div>
-          ) : null}
-          {/* eslint-disable-next-line @next/next/no-img-element */}
-          <img
-            src={imageUrl}
-            alt={artistProfile?.name || "Artist"}
-            className={`size-full object-cover transition-opacity duration-300 ${
-              imgLoaded ? "opacity-100" : "opacity-0"
-            }`}
-            onLoad={() => setImgLoaded(true)}
-            onError={() => setImgErrored(true)}
-          />
-        </div>
+      {imageUrl ? (
+        <ArtistAvatar
+          key={imageUrl}
+          imageUrl={imageUrl}
+          name={artistProfile?.name || "Artist"}
+        />
       ) : (
         <div className="flex size-16 shrink-0 items-center justify-center rounded-xl bg-muted text-muted-foreground shadow-sm">
           <User className="size-6 opacity-60" />
