+++ components/VercelChat/tools/getPlatformVisual.ts
